requirements.txt
opencv-python
pysimplegui
torch
torchvision
pillow
